requirements.txt
# Web Framework
flask==3.0.0
gunicorn==21.2.0

# Slack SDK
slack-bolt==1.18.1
slack-sdk>=3.25.0

# Image Processing
Pillow==10.1.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7  # For MySQL SSL

# Datadog APM
ddtrace==2.3.0
wrapt>=1.14.0

# HTTP Requests
requests>=2.31.0

# Environment
python-dotenv==1.0.0
